Refactor gravity simulation into multiple modules and remove bus (replaced with signal on controller)
