+++ godot/project.godot
@@ -15,10 +15,6 @@ run/main_scene="res://src/main.tscn"
 config/features=PackedStringArray("4.3", "C#", "Forward Plus")
 config/icon="res://icon.svg"
 
-[autoload]
-
-Bus="*res://src/physics/bus.gd"
-
 [dotnet]
 
 project/assembly_name="Kandidat"
