--- godot/src/physics/bus.gd
@@ -1,4 +1,0 @@
-@tool
-extends Node
-
-signal UPDATE_TRAJECTORY
