+++ rust_gdext/src/physics/gravity/mod.rs
@@ -1,0 +1,3 @@
+mod body;
+mod controller;
+mod trajectories;
